requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.0
lxml==4.9.3
beautifulsoup4==4.12.2
